Temporary handling of Disconnecting / Reconnecting
If a player disconnects from a board, end the game and notify the other players in the same group that the game is over.

If a player reconnects: Reuse the old player labels and data, but otherwise treat the reconnecting player as a new player.

Please note that reconnecting players are abstracted away by the GameServer.