Start games with a timer
The board queue is now consumed based on a timer instead of automatically starting games as soon as the queue is filled. Hopefully this will prevent issues where a player disconnects while the game is starting.

Added safe guard that deletes boards that were already added to the scene tree if a player disconnects during the creation of the boards